book: Update installation chapter to describe dCache 1.9 with Chimera
This patch updates the installation chapter of the book to describe how 
to install dCache 1.9 with Chimera. Since this is our recommended setup, 
it makes sense to describe it in the book. I have removed information 
about older versions of dCache, about upgrading from dCache 1.6, and 
about PNFS. The former two because they are no longer current, the 
latter because it is not the recommended setup for dCache.

I suggest that we in the near future update default values in 
node_config etc to make the deployment with Chimera simpler.

Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/355/