dCacheBook: remove need for absolute CSS reference
There is an issue with chunked HTML output in different directories and CSS.  Chunked HTML 
output is when the book's contents is split over more than one web-page, as opposed to having 
the entire book's content on a single page.  We include a link to the page's CSS, as is 
normal, using DocBook's ability to specify one (or more) style sheets to include.

There is a problem with how the standard DocBook stylesheets treat the CSS-stylesheet URIs.  
The URIs are inserted as string literals.  If the chunked output is split into different 
directories then this will only work if the CSS URI is an absolute URI (starting with "/").

Absolute URIs are inconvenient as one cannot easily move the page to another location.  They 
are also bad because one cannot deploy a test-instance with local CSS changes.

This patch fixes this.  It adds the necessary double-dots in front of the CSS name to allow 
the web-browser to correctly locate the CSS file.

Acked-by: Owen
RB: http://rb.dcache.org/r/1242