add support for curly apostrophes
The keyboards we all use can emit only a single form of the apostrophe.  This is usually mapped to 
the straight typewriter-style apostrophe (U+0027 
http://www.fileformat.info/info/unicode/char/0027/index.htm ).  However, there is an ambiguity 
between this character and the form as seem in most written text: the "curly" apostrophe (U+2019 
http://www.fileformat.info/info/unicode/char/2019/index.htm ) also known as the Right Single 
Quotation Mark (RSQM).

This patch adds support for the RSQM.  Any apostrophe character in the text of a paragraph or title 
will be mapped to the RSQM.  This leaves any apostrophe in computer output or program listing 
unchanged.

RB: http://rb.dcache.org/r/1552/
Acked-by: Gerd