Fix broken commit of patch 2022
Patch 2022:

http://rb.dcache.org/r/2022/

was committed to 1.9.7, 1.9.8, 1.9.9 and trunk as SVN commits:

 1.9.7: 294
 1.9.8: 293
 1.9.9: 292
 Trunk: 291

For Trunk, 1.9.7, and 1.9.9, the changes to install.xml was the same and the file continued 
to be identical on these branches.  However, the patch applied to branch 1.9.8 is 
(inexplicably) different.  After applying the patch, the following difference is present:

--- install.xml (.../1.9.7/install.xml) (revision 300)
+++ install.xml (.../1.9.8/install.xml) (revision 300)
@@ -669,7 +669,7 @@
         </note>
 
         <para>
-          All configured components can now be started with
+          All configured components can now be starting with
           <userinput>dcache start</userinput>, for example:
         </para>
 
This is consistent with the patch not updating the line starting "all configured 
components"; however, it is unclear how this could happen.

This patch rectifies this problem by applying the reverse of the above patch.  After this, 
the install.xml file is consistent between 1.9.8 and 1.9.7, 1.9.9 and trunk.