book: simplify installation of dCache
After changing chimera-config.xml in http://rb.dcache.org/r/2555/ the documentation needs to be changed as well.

In the dCache Book in the installation chapter there used to be a section, where the user had to change the user in the chimera-config.xml from 'postgres' to 'chimera'.
The user does not need to edit chimera-config.xml anymore.

Acked-by: Jan
Patch: http://rb.dcache.org/r/2560/
Target: trunk