book: Removing dependency of version in install.xml
Whenever the dCache version is mentioned in the book it had to be changed everywhere. 
Now the dCache version needs to be changed only in the shared-entities.xml after being defined as a new entity.

Substituted 1.9.8-1 by &dcache-version; and added new entity in shared-entities.xml.
 
Acked-by: Paul, Tanja
Patch: http://rb.dcache.org/r/2578/
Target: trunk