Add missing kerberos entity
Commit 427 mistakenly omitted to update the shared-entities.xml file with 
the definition of the kerberos protocol entity.  This update was included 
in the ReviewBoard patch so I believe the change in shared-entities.xml was 
mistakenly not committed.

This patch fixes this mistake by committing the change that was mistakenly 
omitted.  It fixes the problem of the kerberos entity not being defined.