Book: SRM parameters in dcache.conf
In Ticket http://rt.dcache.org/Ticket/Display.html?id=5876 it is
pointed out that an exception will be thrown if the parameters
'CUSTODIAL', 'NEARLINE', ... are not written in capital letters.  The
importance of using capital letters is now pointed out in the
documentation.

Acked-by: Paul
Patch: http://rb.dcache.org/r/2905/#review5593
Target: 1.9.10