book: fix subtitle of 1.9.5 book
The dCache Book is version-specific.  The version for dCache 1.9.5 
currently has the subtitle "for versions prior to 1.9.7-1" which is 
confusing, but necessary while we supported 1.9.5 and 1.9.6.

Given that we no longer support any version of dCache with this 
version of the book other than 1.9.5, we can change the title to the 
much simpler "for 1.9.5-series" which is in keeping with the subtitle 
for the other versions of the dCache book.

Target: 1.9.5
Patch: http://rb.dcache.org/r/3333/
Acked-by: Gerd Behrmann