book: add support for building opt- and FHS- versions of the Book
This patch adjusts the Book building process so it supports building an
"/opt/d-cache"-specific and "FHS"-specific versions of the book.  The
different output is directed using the "profiling" feature of DocBook.
With DocBook, "profiling" has a slightly different meaning from
programming; it refers to filtering out some content to provide
different variations from the same source material.

At the fundamental level, this patch works by introducing a new
profiling attribute 'layout'.  Any element with this attribute is only
displayed if layout parameter has that attribute's value.  The values
'fhs' and 'opt' are used for the two layouts.  Here's an example:

  <para>
    A generic paragraph that appears in FHS and /opt/d-cache
    versions
  </para>

  <para layout='fhs'>A FHS-specific paragraph</para>

  <para layout='opt'>An /opt/d-cache-specific paragraph</para>

The bulk of this patch is updating the Book so that absolute paths
(/opt/d-cache/...) are updated to show /opt/d-cache/... for the
/opt/d-cache case (layout='opt') and a different path for the FHS case
(layout='fhs').  This is achieved with the 'phrase'-markup.  The
'phrase' markup is a generic way of grouping in-line text together
(cannot use it with block-formatted content).  The 'phrase'-markup
carries no semantics, so it doesn't alter the output.  It is also
available within (almost) any element.

Here's an example of using phrase to parametrise a path.

  <phrase layout='opt'>/opt/d-cache/etc</phrase><phrase layout='fhs'>/etc/dcach$

This expands to "<phrase 
layout='opt'>/opt/d-cache/etc</phrase>/dcache.conf" for the /opt/d-cache
layout and "<phrase layout='fhs'>/etc/dcache</phrase>/dcache.conf" for
the FHS-layout.  Since phrase carries no additional semantics, this is
equivalent to "/opt/d-cache/etc/dcache.conf" and
"/etc/dcache/dcache.conf" respectively.

Path expansions, like the above example, are common within the Book.  
Since the mappings tend to follow a fixed form, several entities are
defined (in shared-entities.xml) that make these paths easier to write
and read.

There is a naming convention for these entities that, while not 
guaranteed to avoid clashes, is likely to be "good enough".  The
convention is:

    'path-' <opt label> '-' <FHS label> [ '-s' ]

Where <opt label> and <FHS label> are terse descriptions of the path
when expanded with built with /opt/d-cache and FHS-layouts.  The label
names are taken from the first letter of the directories in the path;
so:

  <phrase layout='opt'>/opt/d-cache/etc</phrase><phrase layout='fhs'>/etc/dcach$

is the entity 'path-ode-ed'.  Using this entity, the above example may
be written as '&path-ode-ed;/dcache.conf'

If '-s' is specified then the entity includes a final slash and if a
label is 'n' then that part is empty (even if the entity name ends
'-s').

For example '&path-odb-n-s;dcache' expands to '/opt/d-cache/bin/dcache'
for the "/opt/d-cache"-layout (ignoring the phrase-element) and to just
'dcache' for the "FHS"-layout case.

Target: 1.9.12
Target: 1.9.13
Target: trunk
Ticket: http://rt.dcache.org/Ticket/Display.html?id=6510