Book: remove outdated information contained in a comment
The book contains a comment describing that information is missing.  
The comment is meant to provide (in loose, note-form) some of the 
information that is missing.

Unfortunately, that information is now so out-of-date that it is 
misleading and will confuse whoever updates the chapter in the future.  
The only part of the information that is still accurate is that more 
documentation is needed.

This patch fixes this by replacing the outdated and potentially 
misleading information with a simple comment that work is needed.

Target: 1.9.5
Target: 1.9.12
Target: 1.9.13
Target: trunk