Book: install: define pools
The description how to define pools in the layout file contained the
line 'waitForFiles=${path}/data'. This is not useful for someone who
tries to install dCache for the first time, as it is unlikely that he
will have created the directory '/data'. Therefore I deleted that
line.  Nevertheless the information about 'waitForFiles' should not be
lost. At the moment I just do not know where it would fit.

Acked-by: Paul
Patch: http://rb.dcache.org/r/3769/
Target: trunk