Book: SRM part1
I took the chapter on SRM, cut it into pieces, reordered them and
filled the gaps.  The section on SRM configuration for experts is not
ready yet.  I did not yet work on the sections: SRMWatch, Configuring
the PostgreSQL Database, Choosing The right hardware and OS for the
SRM node, General SRM Concepts (for developers).

Patch: http://rb.dcache.org/r/4077/
Acked-by: Dmitry Litvintsev
Target: 2.0