fix building of FHS-layout output
The build rules contain an ambiguity due to the FHS layout having an 
extra -fhs in the name; e.g. Book-a4.pdf and Book-fhs-a4.pdf.  There 
ambiguity comes from whether to consider Book-fhs-a4.pdf as the OPT 
layout of Book-fhs.xml or the FHS layout of Book.xml.

To break this symmetry, make will use the first rule defined.

Unfortunately, the first rules are currently the OPT-layout ones.  This 
leads (incorrectly) to builds using Book-fhs.xml as an original input 
that is profiled for opt, resulting in the double-profiled 
Book-fhs-opt.xml.  This double-profile is sufficient to ensure no 
profiled information survives.

This patch fixes the problem by simply reordering the rules.  
Additional text is added to warn that the order is significant.