Book: Chapters for services
In a restructured version of the Book there should be a chapter for each service
.
The chapter about webdav is new but it is based on an old text from Gerd, so might be outdated.

The nfs-chapter should be merged with the config-nfs.xml as there is only one nfs service left, which needs to be configured differently for nfs3 and nfs4.1.

All other chapters also need to be checked as they are already a few months old.

This is just a working version of the Book to be able to work on the services ch
apters. This should not be the final layout.

The files ftp2.txt and xrootd.txt contain text from Gerd about ftp and xrootd.

Target: