add local deploy targets
The Book's Makefile has targets to deploy the book remotely using scp.
This does not allow to deploy the book to a local directory, which would
be
good to have for an automated deploy job using a mounted directory.

This patch adds the targets deploy-local and test-deploy-local that
behave like
the scp versions except that they use cp instead.